APPDEV-1230 Processed review comments
