+++ app/src/main/java/nu/yona/app/ui/YonaActivity.java
@@ -352,22 +352,10 @@ public class YonaActivity extends BaseActivity implements FragmentManager.OnBack
 		if (!NotificationManagerCompat.from(this).areNotificationsEnabled())
 		{
 			Logger.loge("Notifications Disabled", this.getString(R.string.notification_disabled_message));
-			displayErrorAlert(new ErrorMessage(this.getString(R.string.notification_disabled_message)));
+			AppUtils.displayErrorAlert(new ErrorMessage(this.getString(R.string.notification_disabled_message)));
 		}
 	}
 
-	public static void displayErrorAlert(ErrorMessage errorMessage)
-	{
-		AppUtils.runOnUiThread(() -> getGenericAlertDialogWithErrorMessage(errorMessage).show());
-	}
-
-	public static AlertDialog getGenericAlertDialogWithErrorMessage(ErrorMessage errorMessage)
-	{
-		AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
-		alertDialogBuilder.setMessage(errorMessage.getMessage());
-		alertDialogBuilder.setPositiveButton(getActivity().getString(R.string.ok), null);
-		return alertDialogBuilder.create();
-	}
 
 	private void getUserMessages()
 	{
